refactor: 구글맵 로딩 위치 변경
